Update documentation for unitize mixin
According to the unitize.less the documented way of using the unitize mixin is deprecated. Updatet it to the new one.

Actually I don't know wether it is necessary to mention the base value of 16 here. Wouldn't it be better if there was an additional paragraph about the @remScaleFactor and how to change that if necessary?